Bug 632, 657 and 660 Fixes.

+++ src/app/components/workbench/project/project-members/project-members.component.ts
@@ -1,10 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { Component, inject } from '@angular/core';
-import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
+import { NgbDropdownModule, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ProjectService } from '../../../../../helpers/services/project.service';
 import { Subject, take, takeUntil } from 'rxjs';
 import { ToastrService } from 'ngx-toastr';
 import { RouterLink } from '@angular/router';
+import { DeleteModalComponent } from '../../../../dialogs/delete-modal/delete-modal.component';
+import { DummyDataService } from '../../../../../helpers/services/dummy-data.service';
 
 @Component({
   selector: 'app-project-members',
@@ -15,8 +17,10 @@ import { RouterLink } from '@angular/router';
 })
 export class ProjectMembersComponent {
   // Injectors
-  toastr = inject(ToastrService);
-  projectService = inject(ProjectService);
+  private toastr = inject(ToastrService);
+  private modalService = inject(NgbModal);
+  private projectService = inject(ProjectService);
+  private dummyDataService = inject(DummyDataService);
 
   // Variables
   destroy$:Subject<any> = new Subject();
@@ -25,6 +29,7 @@ export class ProjectMembersComponent {
   currentProject:any;
 
   constructor(){
+    this.dummyDataService.loadingState.next(true);
     this.projectService.currentProject$().pipe(takeUntil(this.destroy$)).subscribe((res:any)=>{
       this.currentProject = res;
       if(res) this.getMembers();
@@ -35,13 +40,23 @@ export class ProjectMembersComponent {
     this.projectService.allMembers(this.currentProject.projectId)
     .pipe(take(1), takeUntil(this.destroy$)).subscribe((res:any)=>{
       this.allMembers = res;
+      this.dummyDataService.loadingState.next(false);
     })
   }
 
   removeMember(name:string){
-    this.projectService.removeMember(this.currentProject.projectId, name).pipe(take(1), takeUntil(this.destroy$)).subscribe((res:any)=>{
-      this.getMembers();
-      this.toastr.success('Member Removed!');
+    const modalRef = this.modalService.open(DeleteModalComponent, { animation: true, centered: true, windowClass: 'no-backdrop', backdrop: false });
+    modalRef.componentInstance.modalFor = 'User';
+    modalRef.componentInstance.modalMsg = `Are you sure you want to remove <b>${name}</b> from this project? They will no longer have access.`;
+    modalRef.dismissed.subscribe((res: any) => {
+      if (res){
+        this.dummyDataService.loadingState.next(true);
+        this.projectService.removeMember(this.currentProject.projectId, name).pipe(take(1), takeUntil(this.destroy$)).subscribe((res:any)=>{
+          this.getMembers();
+          this.toastr.success('Member Removed!');
+          this.dummyDataService.loadingState.next(false);
+        })
+      }
     })
   }
